+++ src/neo4j/_sync/io/_pool.py
@@ -50,8 +50,8 @@ from ...api import (
     READ_ACCESS,
 )
 from ...exceptions import (
-    ClientError,
     ConfigurationError,
+    ConnectionAcquisitionTimeoutError,
     DriverError,
     Neo4jError,
     ReadServiceUnavailable,
@@ -409,8 +409,7 @@ class IOPool(abc.ABC):
                     or not self.cond.wait(timeout)
                 ):
                     log.debug("[#0000]  _: <POOL> acquisition timed out")
-                    # TODO: 6.0 - change this to be a DriverError (or subclass)
-                    raise ClientError(
+                    raise ConnectionAcquisitionTimeoutError(
                         "failed to obtain a connection from the pool within "
                         f"{deadline.original_timeout!r}s (timeout)"
                     )